changes in book.ctrl

+++ src/utils/uploadToCloudinary.ts
@@ -7,20 +7,35 @@ interface ICloudinaryUpload {
    url: string;
 }
 
+// Maintain a list of uploaded file names
+const uploadedFiles: string[] = [];
+
 export const uploadFileToCloudinary = async (
    file: Express.Multer.File,
    folder: string
 ): Promise<ICloudinaryUpload> => {
-   const filePath = path.resolve(__dirname, "../../public/data", file.filename);
+   const fileName = file.filename;
+   const filePath = path.resolve(__dirname, "../../public/data", fileName);
+
    try {
-      const format = file.mimetype.split("/")[1];
+      // Check if the file has already been uploaded
+      if (uploadedFiles.includes(fileName)) {
+         throw new Error(`File '${fileName}' has already been uploaded`);
+      }
 
       // Upload file to Cloudinary
       const uploadResult = await cloudinary.uploader.upload(filePath, {
          folder,
-
          quality_analysis: true,
       });
+
+      // Add the uploaded file name to the list
+      uploadedFiles.push(fileName);
+
+      // Delete local file after successful upload
+      if (fs.existsSync(filePath)) {
+         fs.unlinkSync(filePath);
+      }
 
       // Return upload result
       const cloudinaryUpload: ICloudinaryUpload = {
@@ -28,31 +43,19 @@ export const uploadFileToCloudinary = async (
          url: uploadResult.secure_url,
       };
 
-      // Delete local file after successful upload
-      fs.unlink(filePath, (err) => {
-         if (err) {
-            console.error("Error deleting local file:", err);
-         } else {
-            console.log(
-               `Successfully removed local file ${file.filename} after upload to Cloudinary`
-            );
-         }
-      });
-
       return cloudinaryUpload;
    } catch (error) {
       // Delete local file in case of upload failure
-      fs.unlinkSync(filePath);
+      if (fs.existsSync(filePath)) {
+         fs.unlinkSync(filePath);
+      }
 
-      console.error(
-         `Error uploading file to Cloudinary: ${file.filename}`,
-         error
-      );
+      console.error(`Error uploading file to Cloudinary: ${fileName}`, error);
       throw new Error("Failed to upload file to Cloudinary");
    }
 };
 
-export const deleteToCloudinary = async (public_id: string):Promise<void> => {
+export const deleteToCloudinary = async (public_id: string): Promise<void> => {
    try {
       const deleted = await cloudinary.uploader.destroy(public_id);
    } catch (error: any) {
